// SofTec_TMS/NWC.Web/src/app/TMS-Module/Models/vehicle-dto.ts
import { Lookup } from "./common/lookup";
import { StaffListDto } from "./staff-list-dto";

export class VehicleDTO {
  
  Id:string;
  Code: string = "";
  PlateNo: string = "";
  ChassisNo: string = "";
  TransporterTypeName: string;
  TransporterType: string;
  Status: number;
  DeviceCode: string = "";
  SIMCardNo: string = "";
  Group: string;
  Allocation: string;
  Drivers: Array<string>;
  DriverNames: string;
  Staff: Array<StaffListDto> = [];
  ProductionYear: string;
  ProductionYearName: string;
  Branch: string;
  BranchName: string;
  SubBranch: string;
  SubBranchName: string;
  Landmark: string;
  LandmarkName: string;
  Model: string;
  ModelName: string;
  Brand: string;
  BrandName: string;
  Capacity = 0;

  Color: string;
  FuelLitres: string;
  SpeedLimit: string;
  CapacityUnit: string;
  LicenseType: string;
  DriverCode: string;
  CurrentMileage: string;
  HourRate: string;
  EngineNumber: string;
  LicenseNumber: string;
  InsuredBy: string;
  InsuranceNumber: string;
  InsuranceStartDate: string;
  InsuranceEndDate: string;
  Supplier: string;
  FuelCost: string;
  TemperatureSensorMaxValue: string;
  TemperatureSensorMinValue: string;
  Provider: string;
  ProviderName: string;
  LicenseExpiryDate: string;
  ServiceTypeID: string;
  NextExaminationDate: string;
  OwnerIDValue:string;
  OwnerFullNameAR:string;
  ExcelSheetRowId: number;
  ExcelValidation: string;
  IDSource: string;
}
export class VehiclePermitDTO {

  Id: string;
  Code: string = "";
  PlateNo: string = "";
  ChassisNo: string = "";
  TransporterTypeName: string;
  TransporterType: string;
  Status: number;
  DeviceCode: string = "";
  SIMCardNo: string = "";
  Group: string;
  Allocation: string;
  Drivers: Array<string>;
  DriverNames: string;
  Staff: Array<StaffListDto> = [];
  ProductionYear: string;
  ProductionYearName: string;
  Branch: string;
  BranchName: string;
  SubBranch: string;
  SubBranchName: string;
  Landmark: string;
  LandmarkName: string;
  Model: string;
  ModelName: string;
  Brand: string;
  BrandName: string;
  Capacity = 0;

  Color: string;
  FuelLitres: string;
  SpeedLimit: string;
  CapacityUnit: string;
  LicenseType: string;
  DriverCode: string;
  CurrentMileage: string;
  HourRate: string;
  EngineNumber: string;
  LicenseNumber: string;
  InsuredBy: string;
  InsuranceNumber: string;
  InsuranceStartDate: Date;
  InsuranceEndDate: Date;
  Supplier: string;
  FuelCost: string;
  TemperatureSensorMaxValue: string;
  TemperatureSensorMinValue: string;
  Provider: string;
  ProviderName: string;
  LicenseExpiryDate: Date;
  ServiceTypeID: string;
  NextExaminationDate: string;
  OwnerIDValue: string;
  OwnerFullNameAR: string;
  IDSource: string;
}